update: エンドポイント修正
