+++ src/app/api/presets/time/route.ts
@@ -5,15 +5,3 @@ export async function GET() {
     message: "プリセット一覧取得",
   });
 }
-
-export async function PUT() {
-  return NextResponse.json({
-    message: "特定のプリセットを更新",
-  });
-}
-
-export async function DELETE() {
-  return NextResponse.json({
-    message: "特定のプリセットを削除",
-  });
-}
