+++ src/app/api/schedule/new/route.ts
@@ -1,0 +1,7 @@
+import { NextResponse } from "next/server";
+
+export async function POST() {
+  return NextResponse.json({
+    message: "明日の予定をスケジュール",
+  });
+}
